Migrated requests & offers to a single "offers_and_requests" table.
I'll be doing a sort of "two level" STI here, with the "offer_or_request" column determining whether the record is an offer or a request, and thus the base model (Offer or Request). I also introduce a standard "type" field, which will be used for STI so that we can differentiate by type of offer or request.