Add access control table & logic for offers.
This table is necessary to allow us to expand the privacy options around offers
and requests. Moving forward offers and requests can be either public or private,
but when set to private the creator can explicitly make the offer or request
public to their pod and/or one or more of the organizations they are a member
of.