Start work on new dashboard page.
This page will be the new home page (replacing the 'home pod' page) and will present a better overview of relevant activity on JoatU, such as recently posted offers & requests in the user's home pod, a list of upcoming events, and a list of the user's organizations. This page will also provide an interface for explaining what JoatU is and how best to use it.